Update Search2DMatrixII.h
Search 2D Matrix updated. New approach added.
New approach sees the matrix as Binary Search tree with the root to be up-right element. The children of the node are:
 The node below (>= the parent)
 The node on the left (<= the parent)
In contrast to BST, each node may have more than 1 parent.

Similary such matrix could be seen as Min-Heap with root on the up-left corner and Max-Heap with the root ar the bottom-right corner